+++ samples/Demo/Test/Interfaces.cs
@@ -17,9 +17,9 @@ public class Txt : IComponent
 {
 }
 
-public record struct SizeDefinition(int Value, SizeKind Kind);
+public record struct SizeDefinition(float Value, SizeKind Kind);
 
-public record struct ColorDefinition(int Red, int Gree, int Blue, int Transparency);
+public record struct ColorDefinition(float Red, float Gree, float Blue, float Transparency);
 
 public static class StaticStuff
 {
@@ -46,17 +46,17 @@ public class DivDefinition
     public SizeDefinition Width { get; set; } = new(100, SizeKind.Percentage);
     public SizeDefinition Height { get; set; } = new(100, SizeKind.Percentage);
     public ColorDefinition Color { get; set; } = new(0, 0, 0, 255);
-    public int Padding { get; set; }
-    public int Gap { get; set; }
-    public int Radius { get; set; }
-    public int BorderWidth { get; set; }
+    public float Padding { get; set; }
+    public float Gap { get; set; }
+    public float Radius { get; set; }
+    public float BorderWidth { get; set; }
     public Dir Dir { get; set; } = Dir.Column;
     public MAlign MAlign { get; set; } = MAlign.FlexStart;
     public XAlign XAlign { get; set; } = XAlign.FlexStart;
-    public int ComputedHeight { get; set; }
-    public int ComputedWidth { get; set; }
-    public int ComputedX { get; set; }
-    public int ComputedY { get; set; }
+    public float ComputedHeight { get; set; }
+    public float ComputedWidth { get; set; }
+    public float ComputedX { get; set; }
+    public float ComputedY { get; set; }
 }
 
 public abstract class UiComponent
